requirements.txt
gradio
numpy
pillow
torch
bchlib
fastapi
uvicorn
